+++ frontend/src/pages/Admin/CustomCouponDesign/CustomCouponSection.tsx
@@ -1,6 +1,5 @@
 import { CustomCouponSectionContainer, CouponPreviewHeader } from './CustomCouponSection.style';
 import { Spacing, Text } from '../../../style/layout/common';
-import { useState } from 'react';
 import {
   ImageUpLoadInput,
   ImageUpLoadInputLabel,
@@ -8,21 +7,14 @@ import {
   PreviewImageWrapper,
   PreviewLabel,
 } from './style';
+import useUploadImage from '../../../hooks/useUploadImage';
 
 interface CustomCouponSectionProps {
   label: string;
   uploadImageInputId: string;
 }
 const CustomCouponSection = ({ label, uploadImageInputId }: CustomCouponSectionProps) => {
-  const [imgFileUrl, setImgFileUrl] = useState<string>();
-
-  const uploadImageFile = (e: React.ChangeEvent<HTMLInputElement>) => {
-    if (!e.target.files) return;
-    const files = e.target.files[0];
-    // TODO: 해당 로직을 어떻게 저장할지 논의 해보아야 함.ㄴ
-    const imgUrl = URL.createObjectURL(files);
-    setImgFileUrl(imgUrl);
-  };
+  const { imgFileUrl, uploadImageFile } = useUploadImage();
   return (
     <CustomCouponSectionContainer>
       <CouponPreviewHeader>
