+++ frontend/src/pages/Admin/CustomCouponDesign/CustomStampSection.tsx
@@ -1,4 +1,3 @@
-import React, { useState } from 'react';
 import { CustomStampSectionContainer } from './CustomStampSection.style';
 import { Spacing } from '../../../style/layout/common';
 import {
@@ -8,21 +7,15 @@ import {
   PreviewImageWrapper,
   PreviewLabel,
 } from './style';
+import useUploadImage from '../../../hooks/useUploadImage';
 
 interface CustomStampSectionProps {
   label: string;
   uploadImageInputId: string;
 }
 const CustomStampSection = ({ label, uploadImageInputId }: CustomStampSectionProps) => {
-  const [imgFileUrl, setImgFileUrl] = useState<string>();
+  const { imgFileUrl, uploadImageFile } = useUploadImage();
 
-  const uploadImageFile = (e: React.ChangeEvent<HTMLInputElement>) => {
-    if (!e.target.files) return;
-    const files = e.target.files[0];
-    // TODO: 해당 로직을 어떻게 저장할지 논의 해보아야 함.
-    const imgUrl = URL.createObjectURL(files);
-    setImgFileUrl(imgUrl);
-  };
   return (
     <CustomStampSectionContainer>
       <PreviewLabel>{label}</PreviewLabel>
